+++ kamehouse-ui/src/main/webapp/js/tennisworld/scheduled-bookings.js
@@ -29,7 +29,7 @@ class ScheduledBookingService {
         this.#updateView(responseBody, responseCode);
       },
       (responseBody, responseCode, responseDescription, responseHeaders) => {
-        kameHouse.logger.logApiError(responseBody, responseCode, responseDescription, responseHeaders, "Error executing scheduled bookings");
+        kameHouse.logger.logApiError(ScheduledBookingService.#SCHEDULED_BOOKINGS_API_URL, responseBody, responseCode, responseDescription, responseHeaders, "Error executing scheduled bookings");
         kameHouse.plugin.modal.loadingWheelModal.close();
         this.#updateView(responseBody, responseCode);
       });